+++ libs/cardano-ledger-api/test/Test/Cardano/Ledger/Api/State/Imp/QuerySpec.hs
@@ -14,6 +14,7 @@ import Cardano.Ledger.Api.State.Query (
   MemberStatus (..),
   NextEpochChange (..),
   queryCommitteeMembersState,
+  queryDRepDelegations,
   queryDRepState,
  )
 import Cardano.Ledger.BaseTypes
@@ -197,6 +198,39 @@ spec = do
         passEpoch
         expectQueryResult (Set.singleton c1) mempty mempty $
           [(c1, CommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) NoChangeExpected)]
+
+  it "queryDRepDelegationState" $ do
+    (credDrep, delegator, _) <- setupSingleDRep 1_000_000
+
+    kh <- freshKeyHash
+    let cred = KeyHashObj kh
+    _ <- registerStakeCredential cred
+    _ <- delegateToDRep cred (Coin 2_000_000) DRepAlwaysAbstain
+
+    kh2 <- freshKeyHash
+    let cred2 = KeyHashObj kh2
+    _ <- registerStakeCredential cred2
+    _ <- delegateToDRep cred2 (Coin 3_000_000) DRepAlwaysNoConfidence
+
+    let realDRepCred = DRepCredential credDrep
+
+    nes <- getsNES id
+    let abstainDelegations =
+          Map.singleton DRepAlwaysAbstain (Set.fromList [cred])
+        noConfidenceDelegations =
+          Map.singleton DRepAlwaysNoConfidence (Set.fromList [cred2])
+        realDRepDelegations = Map.singleton realDRepCred (Set.fromList [delegator])
+        expectedAllDelegations =
+          realDRepDelegations
+            <> abstainDelegations
+            <> noConfidenceDelegations
+    queryDRepDelegations nes mempty `shouldBe` expectedAllDelegations
+    queryDRepDelegations nes (Set.singleton DRepAlwaysAbstain)
+      `shouldBe` abstainDelegations
+    queryDRepDelegations nes (Set.singleton DRepAlwaysNoConfidence)
+      `shouldBe` noConfidenceDelegations
+    queryDRepDelegations nes (Set.singleton realDRepCred)
+      `shouldBe` realDRepDelegations
 
   it "Committee queries" $ whenPostBootstrap $ do
     (drep, _, _) <- setupSingleDRep 1_000_000